<refactoring>Change name to comment.entity.ts
