+++ back/main_sever/src/posts/comment/comment.module.ts
@@ -6,7 +6,7 @@ import { UsersModule } from 'src/users/users.module';
 
 import { CommentsController } from './comment.controller';
 import { CommentsService } from './comment.service';
-import { CommentModel } from './entities/comments.entity';
+import { CommentModel } from './entities/comment.entity';
 import { CommentUseCases } from './usecase/comment.use-case';
 import { PostsModule } from '../post.module';
