+++ back/main_sever/src/posts/comment/dto/create-comment.dto.ts
@@ -1,4 +1,4 @@
 import { PickType } from "@nestjs/mapped-types";
-import { CommentModel } from "../entities/comments.entity";
+import { CommentModel } from "../entities/comment.entity";
 
 export class CreateCommentDto  extends PickType(CommentModel, [ 'content']){}
