+++ back/main_sever/src/posts/comment/dto/create-response-comment.dto.ts
@@ -1,5 +1,5 @@
 import { PickType } from '@nestjs/mapped-types';
-import { CommentModel } from '../entities/comments.entity';
+import { CommentModel } from '../entities/comment.entity';
 import { IsNotEmpty, IsNumber, Matches } from 'class-validator';
 import { Transform } from 'class-transformer';
