+++ back/main_sever/src/posts/entities/post.entity.ts
@@ -3,7 +3,7 @@ import { BaseModel } from 'src/common/entity/base.entity';
 import { stringValidationMessage } from 'src/common/validation-message/string-validation.message';
 import { Column, Entity, ManyToOne, OneToMany } from 'typeorm';
 import { UserModel } from 'src/users/entities/user.entity';
-import { CommentModel } from '../comment/entities/comments.entity';
+import { CommentModel } from '../comment/entities/comment.entity';
 
 @Entity()
 export class PostModel extends BaseModel {
